Allow all admin UI screens to be usable outside the engine.
This is a re-implementation of 7ea7741 for PR #984 to account for
the changes done in PR #975. It could be argued that we really
just want to namespace the `settings_sub_menu.html.erb` file and
maybe the `_configuration_menu.html.erb` (the general config) as those
are likely to be the only ones extended but for consistency and
to avoid more PR like this I just hit them all.