Use localization for expiration formatting.
Currently the expiration is using the `to_formatted_s` to formating the
date which I believe is a more legacy setup. Using localization `l` is
a better method and how other parts of the app do it[1].

In addition to being more consistent with the rest of the app (reducing
configuration points) the `:short_date` identifier used here is not
even one built into Rails or Spree/Solidus. So this is basically just
falling back on the default formatting:

    > Date::DATE_FORMATS.keys
    => [:short, :long, :db, :number, :long_ordinal, :rfc822, :iso8601]

    > Date.current.to_s :short_date
    => "2017-03-29"

    > Date.current.to_s :butterfinger
    => "2017-03-29"

1. https://github.com/solidusio/solidus/blob/master/backend/app/views/spree/admin/orders/index.html.erb#L151